requirements.txt
numpy==1.16.0
azure-cli-core==2.0.*
azureml-core==1.23.0
dynaconf==3.1.2
notebook
ipykernel